feat(c-api) Extract `engine` and `store` into their own module.
